requirements.txt
torch>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
tqdm>=4.65.0
wandb>=0.15.0
python-dotenv>=1.0.0

